// nextjs/types.ts
export interface Product {
  images: string[];
  id: number;
  category: Category;
  name: string;
  price: number; // Use 'number' for floating-point numbers
  isFeatured: boolean;
  size: Size;
  color: Color;
  imageUrls: string[];
  isArchived: boolean;
  createdAt: string;
  discount: number;
}

export interface Image {
  id: number;
  url: string;
}

export interface Billboard {
  id: number;
  label: string;
  imageUrl: string[];
  createdAt: string;
}

export interface Category {
  id: number;
  name: string;
}

export interface Size {
  id: number;
  name: string;
  value: string;
  createdAt: string;

  // Add more fields as needed
}

export interface Color {
  id: number;
  name: string;
  value: string;
  createdAt: string;
}

export interface OrderItem {
  UserID: number;
  ProductID: number;
  Total: number;
  Quantity: number;
  ID: number | null;
  Product: Product;

  OrderID: number;
  CreatedAt: string;
  DpdatedAt: string;
}